[Gold V] Title: 기차 여행, Time: 656 ms, Memory: 62888 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/10713. 기차 여행/기차 여행.java
@@ -49,7 +49,7 @@ public class Main {
 		}
 		
 		long ans = 0; 
-		int psum = 0;
+		long psum = 0;
 
 		for(int i = 1; i < N; i++) {
 			psum += d[i];
